// frontend/src/api/generated/model/inspectionStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */

export type InspectionStatus = typeof InspectionStatus[keyof typeof InspectionStatus];


export const InspectionStatus = {
  Scheduled: 'Scheduled',
  InProgress: 'InProgress',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;